2016.07.27. Do we need a round for parameters
Not sure why we need a rounding for these values.
In Qt4 version I fixed DataModelInputBinder so values are displayed in format, but actual value is kept in the data model.
We experienced a bug when energy value was rounded and energy change was requested when it was not needed.